wireframes/tools/requirements.txt
Pillow==10.0.0
pytesseract==0.3.10
PyYAML==6.0.1
opencv-python==4.8.0.76
numpy==1.24.3